+++ examples/fluid/hydrostatic_water_column_2d.jl
@@ -64,7 +64,7 @@ boundary_model = BoundaryModelDummyParticles(tank.boundary.density, tank.boundar
                                              boundary_density_calculator,
                                              smoothing_kernel, smoothing_length,
                                              viscosity=viscosity_wall)
-boundary_system = BoundarySPHSystem(tank.boundary, boundary_model, movement=nothing)
+boundary_system = WallBoundarySystem(tank.boundary, boundary_model, movement=nothing)
 
 # ==========================================================================================
 # ==== Simulation