+++ src/schemes/fluid/weakly_compressible_sph/rhs.jl
@@ -47,7 +47,7 @@ function interact!(dv, v_particle_system, u_particle_system,
         # The following call is equivalent to
         #     `p_a = current_pressure(v_particle_system, particle_system, particle)`
         #     `p_b = current_pressure(v_neighbor_system, neighbor_system, neighbor)`
-        # Only when the neighbor system is a `BoundarySPHSystem` or a `TotalLagrangianSPHSystem`
+        # Only when the neighbor system is a `WallBoundarySystem` or a `TotalLagrangianSPHSystem`
         # with the boundary model `PressureMirroring`, this will return `p_b = p_a`, which is
         # the pressure of the fluid particle.
         p_a,
@@ -160,7 +160,7 @@ end
 
 @inline function particle_neighbor_pressure(v_particle_system, v_neighbor_system,
                                             particle_system,
-                                            neighbor_system::BoundarySPHSystem{<:BoundaryModelDummyParticles{PressureMirroring}},
+                                            neighbor_system::WallBoundarySystem{<:BoundaryModelDummyParticles{PressureMirroring}},
                                             particle, neighbor)
     p_a = current_pressure(v_particle_system, particle_system, particle)