+++ test/schemes/boundary/open_boundary/characteristic_variables.jl
@@ -63,7 +63,7 @@
                                                            density_calculator=ContinuityDensity(),
                                                            smoothing_length, sound_speed)
 
-                boundary_system = OpenBoundarySPHSystem(boundary_zone;
+                boundary_system = OpenBoundarySystem(boundary_zone;
                                                         fluid_system, buffer_size=0,
                                                         boundary_model=BoundaryModelCharacteristicsLastiwka())