+++ NEWS.md
@@ -8,8 +8,8 @@ used in the Julia ecosystem. Notable changes will be documented in this file for
 
 ### API Changes
 
-- API for `OpenBoundarySPHSystem` and `BoundaryZone` changed.
-  It is now possible to pass multiple `BoundaryZone`s to a single `OpenBoundarySPHSystem`.
+- API for `OpenBoundarySystem` and `BoundaryZone` changed.
+  It is now possible to pass multiple `BoundaryZone`s to a single `OpenBoundarySystem`.
   Reference values are now assigned individually to each `BoundaryZone`. (#866)
 
 - The argument of `TransportVelocityAdami` is now a keyword argument.